requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
Pillow==10.0.1
urllib3==2.0.6
